Changed the version of the guava plug-in to 15.0.
This commit fixes issue #31.

The problem that Eclipse picked up an older version of Guava that caused a
`NoSuchMethodError` exception because the older version didn't
Iterables.tryFind(...) method. As a result, Eclipse failed to create the
CompatibilityView object. The lesson here is that I should've paid closer
attention to all exceptions not only the first one in the error log because the
root cause was captured by the second exception.