Fixed #35.
What was happening is the following. The system is searching for the error text "addEdges(numvert)".
It searches for the string using the Pattern class which accepts regular expressions.
Since parentheses means grouping in regular expressions, it is technically searching for "addEdgesnumvert"
instead and couldn't find it. To fix this, I added Pattern.quote(...) in order to escape the parentheses.