Made the marker locator more precise.
Instead of the using the narrow piece of code reported in the error messages of the type checker, I used the enclosing line of the reported piece of code to locate the marker in the primary project more precisely. The enclosing line results in more precise marker location because it's longer and is more likely to be unique in the enclosing compilation unit.

issue #62